third_party/WebKit/LayoutTests/bluetooth/requestDevice/request-from-sandboxed-iframe.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<body>
  <script>
  "use strict";
  async_test(test => {
    window.onmessage = messageEvent => test.step(() => {
      if (messageEvent.data === 'Ready') {
        let iframe = document.querySelector('iframe');
        callWithTrustedClick(() => {
          iframe.contentWindow.postMessage('Go', '*');
        });
      } else {
        assert_equals(messageEvent.data, 'SecurityError: requestDevice() ' +
                                         'called from cross-origin iframe.');
        test.done();
      }
    });
    setBluetoothFakeAdapter('HeartRateAdapter')
      .then(() => {
        let iframe = document.createElement('iframe');
        iframe.sandbox.add('allow-scripts');
        iframe.src = '../../resources/bluetooth/requestDevice-in-iframe.html';
        document.body.appendChild(iframe);
      });
  }, 'Request device from a unique origin. Should reject with SecurityError.');
  </script>
</body>
